docs: update readme and refactor config
